# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_acerowl_lockin_backend_internal_models.Lesson:
    properties:
      content:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      index:
        type: integer
      nodeId:
        type: string
      quizzes:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Question'
        type: array
      status:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Status'
      title:
        type: string
    type: object
...
  github_com_acerowl_lockin_backend_internal_models.Module:
    properties:
      concept_tags:
        items:
          type: string
        type: array
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      index:
        type: integer
      lessons:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Lesson'
        type: array
      parentNodeId:
        type: string
      status:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Status'
      title:
        type: string
      topicId:
        type: string
    type: object
  github_com_acerowl_lockin_backend_internal_models.Option:
    properties:
      explanation:
        type: string
      id:
        type: string
      index:
        type: integer
      is_correct:
        type: boolean
      label:
        type: string
      questionId:
        type: string
    type: object
  github_com_acerowl_lockin_backend_internal_models.Question:
    properties:
      answer:
        type: string
      concept_tags:
        items:
          type: string
        type: array
      createdAt:
        type: string
      explanation:
        type: string
      id:
        type: string
      index:
        type: integer
      lessonId:
        type: string
      nodeId:
        type: string
      options:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Option'
        type: array
      question:
        type: string
      type:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.QuestionType'
    type: object
  github_com_acerowl_lockin_backend_internal_models.QuestionType:
    enum:
    - mcq
    - true_false
    - fill_blank
    - short_answer
    - speech
    type: string
    x-enum-varnames:
    - MCQ
    - TrueFalse
    - FillBlank
    - ShortAnswer
    - Speech
...
  github_com_acerowl_lockin_backend_internal_models.Status:
    enum:
    - locked
    - in-progress
    - completed
    type: string
    x-enum-varnames:
    - StatusLocked
    - StatusInProgress
    - StatusCompleted
  github_com_acerowl_lockin_backend_internal_models.Topic:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      remark:
        type: string
      status:
        type: string
      tier:
        type: integer
      title:
        type: string
      userId:
        type: string
    type: object
  github_com_acerowl_lockin_backend_internal_models.TopicRoadmap:
    properties:
      id:
        type: string
      modules:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Module'
        type: array
      sessionsCompleted:
        type: integer
      tier:
        type: integer
      title:
        type: string
      totalTimeSeconds:
        type: number
    type: object
...
  internal_handlers.AssessmentResponse:
    properties:
      questions:
        properties:
          data:
            items:
              $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Question'
            type: array
        type: object
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.CreateTopicData:
    properties:
      status:
        example: generating
        type: string
      topic:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Topic'
    type: object
  internal_handlers.CreateTopicReq:
    properties:
      familiarity_level:
        description: beginner, intermediate, advanced
        example: beginner
        type: string
      title:
        example: Quantum Mechanics
        type: string
    type: object
  internal_handlers.CreateTopicResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.CreateTopicData'
      success:
        example: true
        type: boolean
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  internal_handlers.RoadmapResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.TopicRoadmap'
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.TopicListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Topic'
        type: array
      success:
        example: true
        type: boolean
    type: object
  internal_handlers.TopicResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Topic'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: LockIn Support
    url: https://github.com/acerowl/lockin
  description: Adaptive learning backend — quiz generation, sessions, and progress
    tracking.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: LockIn API
  version: "1.0"
paths:
...
  /topics:
    get:
      description: Returns all topics for the authenticated user, ordered by creation
        date (newest first)
      produces:
      - application/json
      responses:
        "200":
          description: Array of user topics
          schema:
            $ref: '#/definitions/internal_handlers.TopicListResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List topics
      tags:
      - topics
    post:
      consumes:
      - application/json
      description: Creates a new learning topic. An initial tier is assigned based
        on familiarity_level (beginner=1, intermediate=4, advanced=7). After AI generates
        the roadmap, the tier is updated based on the AI response.
      parameters:
      - description: Topic payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateTopicReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created topic with generation status
          schema:
            $ref: '#/definitions/internal_handlers.CreateTopicResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a topic
      tags:
      - topics
  /topics/{id}:
    get:
      description: Fetches a single topic by ID for the authenticated user
      parameters:
      - description: Topic UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Topic details
          schema:
            $ref: '#/definitions/internal_handlers.TopicResponse'
        "404":
          description: Topic not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a topic
      tags:
      - topics
...
  /topics/assessment:
    post:
      description: Generate question for assessment of user's knowledge on the topic
      produces:
      - application/json
      responses:
        "200":
          description: Assessment questions
          schema:
            $ref: '#/definitions/internal_handlers.AssessmentResponse'
        "404":
          description: Error generating topic assessment
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate assessment question for the topic
      tags:
      - topics
  /topics/roadmap/{id}:
    get:
      description: Fetches the learning roadmap (modules and lessons) for a topic
      parameters:
      - description: Topic UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Roadmap with modules
          schema:
            $ref: '#/definitions/internal_handlers.RoadmapResponse'
        "404":
          description: Topic or roadmap not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get roadmap of a topic
      tags:
      - topics
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"